LayoutTests/css3/calc/table-border-spacing.html

<!DOCTYPE html>
<script src="../../fast/js/resources/js-test-pre.js"></script>
<style>
    td { 
        padding: 0; 
        width: 50px; 
        height: 60px; 
        background-color: blue; 
        border: 1px solid black 
    }
    table { 
        border-spacing: -webkit-calc(10px * 2 + 5px); 
        border: 3px solid black; 
    }
</style>
<table id="stimulus">
    <tr>
        <td></td>
        <td></td>
        <td></td>
        <td></td>
        <td></td>
    </tr>
    <tr>
        <td></td>
        <td></td>
        <td></td>
        <td></td>
        <td></td>
    </tr>
</table>
<script>
    var innerBorder = 1;
    var outerBorder = 3;
    var innerWidth = 50 + 2 * innerBorder;
    var innerHeight = 60 + 2 * innerBorder;
    var borderSpacing = 25;
    var rows = 2;
    var columns = 5;

    var expectedWidth = innerWidth * columns + borderSpacing * (columns + 1) + 2 * outerBorder;
    var expectedHeight = innerHeight * rows + borderSpacing * (rows + 1) + 2 * outerBorder;

    shouldEvaluateTo('document.getElementById("stimulus").offsetWidth', expectedWidth);
    shouldEvaluateTo('document.getElementById("stimulus").offsetHeight', expectedHeight);
</script>    
<script src="../../fast/js/resources/js-test-post.js"></script>
